Move MrDlibPreferences to GuiPreferences
