+++ jabgui/src/main/java/org/jabref/gui/preferences/GuiPreferences.java
@@ -17,6 +17,7 @@ import org.jabref.gui.mergeentries.MergeDialogPreferences;
 import org.jabref.gui.newentry.NewEntryPreferences;
 import org.jabref.gui.preview.PreviewPreferences;
 import org.jabref.gui.specialfields.SpecialFieldsPreferences;
+import org.jabref.logic.importer.fetcher.MrDlibPreferences;
 import org.jabref.logic.preferences.CliPreferences;
 import org.jabref.logic.push.PushToApplicationPreferences;
 
@@ -60,4 +61,6 @@ public interface GuiPreferences extends CliPreferences {
     NewEntryPreferences getNewEntryPreferences();
 
     DonationPreferences getDonationPreferences();
+
+    MrDlibPreferences getMrDlibPreferences();
 }
