+++ jabgui/src/main/java/org/jabref/gui/preferences/JabRefGuiPreferences.java
@@ -50,6 +50,7 @@ import org.jabref.logic.exporter.BibDatabaseWriter;
 import org.jabref.logic.exporter.SelfContainedSaveConfiguration;
 import org.jabref.logic.externalfiles.DateRange;
 import org.jabref.logic.externalfiles.ExternalFileSorter;
+import org.jabref.logic.importer.fetcher.MrDlibPreferences;
 import org.jabref.logic.importer.fetcher.citation.CitationCountFetcherType;
 import org.jabref.logic.importer.fetcher.citation.CitationFetcherType;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
@@ -230,6 +231,13 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
     private static final String CITATION_COUNT_FETCHER_TYPE = "citationCountFetcherType";
     // endregion
 
+    // region MrDlibPreferences
+    private static final String MRDLIB_ACCEPT_RECOMMENDATIONS = "acceptRecommendations";
+    private static final String MRDLIB_SEND_LANGUAGE_DATA = "sendLanguageData";
+    private static final String MRDLIB_SEND_OS_DATA = "sendOSData";
+    private static final String MRDLIB_SEND_TIMEZONE_DATA = "sendTimezoneData";
+    // endregion
+
     private static JabRefGuiPreferences singleton;
 
     private EntryEditorPreferences entryEditorPreferences;
@@ -251,6 +259,7 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
     private CopyToPreferences copyToPreferences;
     private NewEntryPreferences newEntryPreferences;
     private DonationPreferences donationPreferences;
+    private MrDlibPreferences mrDlibPreferences;
 
     /// @deprecated Never ever add a call to this method. There should be only one caller.
     /// All other usages should get the preferences passed (or injected).
@@ -288,6 +297,7 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
                 getLayoutFormatterPreferences(),
                 Injector.instantiateModelOrService(JournalAbbreviationRepository.class),
                 Injector.instantiateModelOrService(BibEntryTypesManager.class)));
+        getMrDlibPreferences().setAll(MrDlibPreferences.getDefault());
     }
 
     @Override
@@ -313,6 +323,7 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
         getSidePanePreferences().setAll(getSidePanePreferencesFromBackingStore(getSidePanePreferences()));
         getNameDisplayPreferences().setAll(getNameDisplayPreferencesFromBackingStore(getNameDisplayPreferences()));
         getPreviewPreferences().setAll(getPreviewPreferencesFromBackingStore(getPreviewPreferences()));
+        getMrDlibPreferences().setAll(getMrDlibPreferencesFromBackingStore(getMrDlibPreferences()));
     }
 
     // region CopyToPreferences
@@ -1177,6 +1188,32 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
         return new DonationPreferences(
                 getBoolean(DONATION_NEVER_SHOW, defaults.isNeverShowAgain()),
                 getInt(DONATION_LAST_SHOWN_EPOCH_DAY, defaults.getLastShownEpochDay()));
+    }
+    // endregion
+
+    // region MrDlibPreferences
+    @Override
+    public MrDlibPreferences getMrDlibPreferences() {
+        if (mrDlibPreferences != null) {
+            return mrDlibPreferences;
+        }
+
+        mrDlibPreferences = getMrDlibPreferencesFromBackingStore(MrDlibPreferences.getDefault());
+
+        EasyBind.listen(mrDlibPreferences.acceptRecommendationsProperty(), (_, _, newValue) -> putBoolean(MRDLIB_ACCEPT_RECOMMENDATIONS, newValue));
+        EasyBind.listen(mrDlibPreferences.sendLanguageProperty(), (_, _, newValue) -> putBoolean(MRDLIB_SEND_LANGUAGE_DATA, newValue));
+        EasyBind.listen(mrDlibPreferences.sendOsProperty(), (_, _, newValue) -> putBoolean(MRDLIB_SEND_OS_DATA, newValue));
+        EasyBind.listen(mrDlibPreferences.sendTimezoneProperty(), (_, _, newValue) -> putBoolean(MRDLIB_SEND_TIMEZONE_DATA, newValue));
+
+        return mrDlibPreferences;
+    }
+
+    private MrDlibPreferences getMrDlibPreferencesFromBackingStore(MrDlibPreferences defaults) {
+        return new MrDlibPreferences(
+                getBoolean(MRDLIB_ACCEPT_RECOMMENDATIONS, defaults.shouldAcceptRecommendations()),
+                getBoolean(MRDLIB_SEND_LANGUAGE_DATA, defaults.shouldSendLanguage()),
+                getBoolean(MRDLIB_SEND_OS_DATA, defaults.shouldSendOs()),
+                getBoolean(MRDLIB_SEND_TIMEZONE_DATA, defaults.shouldSendTimezone()));
     }
     // endregion
